new challenge completed - using_as

+++ exercises/conversions/using_as.rs
@@ -5,24 +5,22 @@
 // The goal is to make sure that the division does not fail to compile
 // and returns the proper type.
 
-// I AM NOT DONE
-
 fn average(values: &[f64]) -> f64 {
-    let total = values.iter().fold(0.0, |a, b| a + b);
-    total / values.len()
+  let total = values.iter().fold(0.0, |a, b| a + b);
+  total / values.len() as f64
 }
 
 fn main() {
-    let values = [3.5, 0.3, 13.0, 11.7];
-    println!("{}", average(&values));
+  let values = [3.5, 0.3, 13.0, 11.7];
+  println!("{}", average(&values));
 }
 
 #[cfg(test)]
 mod tests {
-    use super::*;
+  use super::*;
 
-    #[test]
-    fn returns_proper_type_and_value() {
-        assert_eq!(average(&[3.5, 0.3, 13.0, 11.7]), 7.125);
-    }
+  #[test]
+  fn returns_proper_type_and_value() {
+    assert_eq!(average(&[3.5, 0.3, 13.0, 11.7]), 7.125);
+  }
 }
